fix(esp32-3.x): repair enum-migration ripple that broke the Arduino-ESP32 3.x build
The Arduino Build (Arduino-ESP32 3.x core) job was red from residual enum-class
migration fallout that the native suite (looser host GCC) did not catch. Root-cause
fixes, propagating the enum type and casting only at genuine wire/format boundaries
(no enum-silencing casts):

- mqtt.cpp: switch on a uint8_t wire byte with enum-class MqttType case labels ->
  convert the scrutinee once, `switch ((MqttType)type)`.
- preempt_queue.cpp: mangled cast `static_cast<detws_pq_lane>(uintptr_t) arg`
  (uintptr_t is a type, not an expression) -> `static_cast<detws_pq_lane>(reinterpret_cast<uintptr_t>(arg))`.
- oauth2.cpp: ternary mixing an int HTTP status with a DetwsOAuth2Result member in an
  int-returning helper -> cast the member `(int)` (matches the other returns).
- examples 50.OidcAuth / 69.EthernetDnc: `int rc = <enum-class-returning call>` ->
  keep rc typed as the result enum; cast only at the `%d` printf.
- examples 16.ZWaveGateway / 17.ZigbeeGateway / 62.Ld2410: qualify constants that moved
  into namespacing structs (Zwave::ZWAVE_ACK, Ash::ASH_RST, Ld2410State::LD2410_STATE_*).

Verified on the RPi: affected native envs from test/dep_graph.json
(native_mqtt 22, native_oauth2 9, native_preempt_queue 12) all green, and all 8 example
sketches compile clean on the Arduino-ESP32 3.x core (esp32:esp32@3.3.10, arduino-cli).

+++ src/services/oauth2/oauth2.cpp
@@ -172,8 +172,9 @@ int post_and_parse(Oauth2Ctx &c, const char *token_url, int body_len, DetwsOAuth
         memcpy(c.resp, r.body, k);
     c.resp[k] = '\0';
     if (!detws_oauth2_parse_token_response(c.resp, out))
-        return st >= 400 ? st
-                         : DetwsOAuth2Result::DETWS_OAUTH2_ERR_RESPONSE; // surface the provider's 4xx, else generic
+        return st >= 400
+                   ? st
+                   : (int)DetwsOAuth2Result::DETWS_OAUTH2_ERR_RESPONSE; // surface the provider's 4xx, else generic
     return st;
 }
 } // namespace
